Project/djnagor_Rest_Framework/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-allauth
dj-rest-auth
requests-oauthlib
requests
drf-spectacular
celery
redis
